fix planet arrival freeze and crashes.
