+++ NebulaPatcher/Patches/Dynamic/PlanetModelingManager_Patch.cs
@@ -4,7 +4,6 @@ using NebulaModel.Packets.Planet;
 using NebulaWorld;
 using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
 
 namespace NebulaPatcher.Patches.Dynamic
 {
@@ -27,6 +26,16 @@ namespace NebulaPatcher.Patches.Dynamic
 
             // They appear to have conveniently left this flag in for us, but they don't use it anywhere
             planet.factoryLoading = true;
+
+            // do this here to match the patch in GPUInstancingManager_Patch.cs
+            // as we sync entity placement in realtime when players change something
+            // we only need to request the full factory if we never received it before
+            if (planet.factory != null)
+            {
+                AccessTools.Field(typeof(PlanetModelingManager), "currentFactingPlanet").SetValue(null, planet);
+                AccessTools.Field(typeof(PlanetModelingManager), "currentFactingStage").SetValue(null, 0);
+                return false;
+            }
 
             // Request factory
             Log.Info($"Requested factory for planet {planet.name} (ID: {planet.id}) from host");
@@ -91,92 +100,6 @@ namespace NebulaPatcher.Patches.Dynamic
                 {
                     LocalPlayer.SendPacket(new PlanetDataRequest(planetsToRequest.ToArray()));
                 }
-            }
-        }
-    }
-    [HarmonyPatch(typeof(PlanetModelingManager), "LoadingPlanetFactoryMain")]
-    public class PlanetModelingManager_Patch2
-    {
-        public static bool Prefix(PlanetData planet)
-        {
-            if (!SimulatedWorld.Initialized || LocalPlayer.IsMasterClient)
-            {
-                return true;
-            }
-
-            //if we are the client we always need to call GetOrCreateFactory() as this is where we handle the FactoryData received from the server
-            // NOTE: currentFactingStage is a private field so i need to use the refstub for now
-            int currentFactingStage = (int)AccessTools.Field(typeof(PlanetModelingManager), "currentFactingStage").GetValue(null);
-            if (planet.factory != null && currentFactingStage == 0)
-            {
-                GameMain.data.GetOrCreateFactory(planet);
-            }
-            return true;
-        }
-
-        // NOTE: this is part of the weird planet movement fix, see ArrivePlanet() patch for more information
-        [HarmonyPatch(typeof(GameData), "OnActivePlanetLoaded")]
-        class GameData_Patch2
-        {
-            public static bool Prefix(GameData __instance, PlanetData planet)
-            {
-                if (LocalPlayer.IsMasterClient)
-                {
-                    return true;
-                }
-                if (planet != null)
-                {
-                    if (planet.factoryLoaded)
-                    {
-                        __instance.OnActivePlanetFactoryLoaded(planet);
-                    }
-                    else
-                    {
-                        planet.LoadFactory();
-                        planet.onFactoryLoaded += __instance.OnActivePlanetFactoryLoaded;
-                    }
-                }
-                planet.onLoaded -= __instance.OnActivePlanetLoaded;
-                return false;
-            }
-        }
-        // NOTE: this is part of the weird planet movement fix, see ArrivePlanet() patch for more information
-        [HarmonyPatch(typeof(GameData), "OnActivePlanetFactoryLoaded")]
-        class GameData_Patch3
-        {
-            public static bool Prefix(GameData __instance, PlanetData planet)
-            {
-                if (LocalPlayer.IsMasterClient)
-                {
-                    return true;
-                }
-                if (planet != null)
-                {
-                    if (GameMain.gameTick == 0L && DSPGame.SkipPrologue)
-                    {
-                        GameData_Patch3_Helper.InitLandingPlace(__instance, planet);
-                    }
-                    // now set localPlanet and planetId
-                    AccessTools.Property(typeof(GameData), "localPlanet").SetValue(GameMain.data, planet, null);
-                    __instance.mainPlayer.planetId = planet.id;
-                }
-                planet.onFactoryLoaded -= __instance.OnActivePlanetFactoryLoaded;
-                return false;
-            }
-        }
-        class GameData_Patch3_Helper
-        {
-            public static void InitLandingPlace(GameData gameData, PlanetData planet)
-            {
-                Vector3 birthPoint = planet.birthPoint;
-                Quaternion quaternion = Maths.SphericalRotation(birthPoint, 0f);
-                gameData.mainPlayer.transform.localPosition = birthPoint;
-                gameData.mainPlayer.transform.localRotation = quaternion;
-                gameData.mainPlayer.transform.localScale = Vector3.one;
-                gameData.mainPlayer.uPosition = (Vector3)planet.uPosition + planet.runtimeRotation * birthPoint;
-                gameData.mainPlayer.uRotation = planet.runtimeRotation * quaternion;
-                gameData.mainPlayer.uVelocity = VectorLF3.zero;
-                gameData.mainPlayer.controller.velocityOnLanding = Vector3.zero;
             }
         }
     }
